requirements.txt
# Core Dependencies
flet>=0.24.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
chromadb>=0.5.0
pydantic>=2.0.0

# LLM Provider SDKs (Phase 1)
openai>=1.0.0
anthropic>=0.18.0

# Agent Frameworks
crewai>=0.80.0
pyautogen>=0.2.0

# Database & Persistence
python-dotenv>=1.0.0
platformdirs>=4.0.0

# Phase 4: Tool Belt Dependencies
unidiff>=0.7.5              # Patch parsing and application
sentence-transformers>=2.2.0  # CPU-based embeddings for RAG
ddgs>=1.0.0                 # DuckDuckGo metasearch for web_search tool (Phase 3)

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
ruff>=0.7.0

# Packaging (for .exe build)
pyinstaller>=6.0.0
